Feat: build hisory model in django and conncect with android multi mode

+++ android/RunUsAndroid/app/src/main/java/MultiMode/UserDistance.java
@@ -7,9 +7,9 @@ public class UserDistance implements Serializable {
 
     MultiModeUser user;
 
-    double distance;
+    float distance;
 
-    public UserDistance(MultiModeUser user, double distance) {
+    public UserDistance(MultiModeUser user, float distance) {
         this.user = user;
         this.distance = distance;
     }
@@ -22,11 +22,11 @@ public class UserDistance implements Serializable {
         this.user = user;
     }
 
-    public double getDistance() {
+    public float getDistance() {
         return distance;
     }
 
-    public void setDistance(double distance) {
+    public void setDistance(float distance) {
         this.distance = distance;
     }
 }
